// packages/workflow/src/schematics/averos-service/schema.ts
/**
 * @license
 * SPDX-License-Identifier: MIT
 *
 * Copyright (c) 2020-2026 Houssemeddine LAOUITI (Wiforge)
 * https://www.wiforge.com
 *
 * This source code is licensed under the MIT license found in the
 * LICENSE file in the root of this repository.
 */

export interface AverosServiceOption {
  // The name of the entity service.
  name: string

  // The name of the target entity.
  ename: string

  // update and assign the service to its entity
  assignToEntity?: boolean

  // The path to create the service.
  path?: string

  // The name of the project.
  project?: string
}